fix: Использование модульного css в компонентах
