+++ react-burger/src/components/app-header/app-header.js
@@ -1,5 +1,7 @@
 import { BurgerIcon, ListIcon, ProfileIcon } from "@ya.praktikum/react-developer-burger-ui-components"
 import { Logo } from "@ya.praktikum/react-developer-burger-ui-components"
+import styles from "./app-header.module.css"
+
 
 const ulStyle = {
   display: "flex",
@@ -18,19 +20,19 @@ const liStyle = {
 const AppHeader = () => {
   return (
     <nav>
-      <ul style={ulStyle}>
-        <li style={liStyle} className={"mt-4 mb-4 ml-1 mb-1 pl-5 pr-5"} key={1}>
+      <ul className={styles.ul}>
+        <li className={`mt-4 mb-4 ml-1 mb-1 pl-5 pr-5 ${styles.li}`} key={1}>
           <BurgerIcon/>
           <span className="text text_type_main-default ml-2 mr-2">Конструктор</span>
         </li>
-        <li style={liStyle} className={"mt-4 mb-4 ml-1 mb-1 pl-5 pr-5"} key={2}>
+        <li className={`mt-4 mb-4 ml-1 mb-1 pl-5 pr-5 ${styles.li}`} key={2}>
           <ListIcon/>
           <span className="text text_type_main-default ml-2 mr-2">Лента заказов</span>
         </li>
-        <li style={liStyle} className={"mt-4 mb-4"} key={3}>
+        <li className={`mt-4 mb-4 ${styles.li}`} key={3}>
           <Logo/>
         </li>
-        <li style={liStyle} className={"mt-4 mb-4 ml-30"} key={4}>
+        <li className={`mt-4 mb-4 ml-30 ${styles.li}`} key={4}>
           <ProfileIcon/>
           <span className="text text_type_main-default ml-2 mr-2">Личный кабинет</span>
         </li>
