+++ react-burger/src/components/app/app.module.css
@@ -1,0 +1,15 @@
+.header {
+    height: 60;
+}
+
+.main {
+    display: flex;
+    justify-content: center;
+    align-content: center;
+}
+
+.section {
+    display: flex;
+    justify-content: flex-start;
+    flex-direction: column;
+}
