+++ react-burger/src/components/burger-ingredients/burger-ingredients.js
@@ -1,6 +1,7 @@
 import { Tab } from "@ya.praktikum/react-developer-burger-ui-components"
 import { CurrencyIcon } from "@ya.praktikum/react-developer-burger-ui-components"
 import data from "../../utils/data"
+import styles from "./burger-ingredients.module.css"
 
 
 const ulStyle = {
@@ -47,25 +48,25 @@ const tabStyle = {
 const BurgerIngredients = () => {
   return (
     <>
-      <div style={tabContainerStyle}>
-        <Tab value="puns" active={true} key={1} style={tabStyle}>
+      <div className={styles.tab_container}>
+        <Tab value="puns" active={true} key={1} className={styles.tab}>
           Булки
         </Tab>
-        <Tab value="sauces" key={2} style={tabStyle}>
+        <Tab value="sauces" key={2} className={styles.tab}>
           Соусы
         </Tab>
-        <Tab value="toppings" key={3} style={tabStyle}>
+        <Tab value="toppings" key={3} className={styles.tab}>
           Начинки
         </Tab>
       </div>
-      <div style={divStyle}>
+      <div className={styles.div}>
         <h2 className="text text_type_main-medium mt-10 mb-6">Булки</h2>
-        <ul style={ulStyle}>
+        <ul className={styles.ul}>
           {data.filter(item => item.type === "bun").map((item, key) => {
               return (
-                <li style={liStyle} key={key}>
+                <li className={styles.li} key={key}>
                   <img src={item.image} alt={item.name}/>
-                  <p style={pricePStyle}>
+                  <p className={styles.price_p}>
                     <span className="text text_type_main-default">20</span>
                     <CurrencyIcon/>
                   </p>
@@ -75,12 +76,12 @@ const BurgerIngredients = () => {
           })}
         </ul>
         <h2 className="text text_type_main-medium mt-10 mb-6">Соусы</h2>
-        <ul style={ulStyle}>
+        <ul className={styles.ul}>
           {data.filter(item => item.type === "sauce").map((item, key) => {
               return (
-                <li style={liStyle} key={key}>
+                <li className={styles.li} key={key}>
                   <img src={item.image} alt={item.name}/>
-                  <p style={pricePStyle}>
+                  <p className={styles.price_p}>
                     <span className="text text_type_main-default">30</span>
                     <CurrencyIcon/>
                   </p>
@@ -90,12 +91,12 @@ const BurgerIngredients = () => {
           })}
         </ul>
         <h2 className="text text_type_main-medium mt-10 mb-6">Начинки</h2>
-        <ul style={ulStyle}>
+        <ul className={styles.ul}>
           {data.filter(item => item.type === "main").map((item, key) => {
               return (
-                <li style={liStyle} key={key}>
+                <li className={styles.li} key={key}>
                   <img src={item.image} alt={item.name}/>
-                  <p style={pricePStyle}>
+                  <p className={styles.price_p}>
                     <span className="text text_type_main-default">40</span>
                     <CurrencyIcon/>
                   </p>
