+++ react-burger/src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,0 +1,40 @@
+.ul {
+    display: grid;
+    list-style: none;
+    flex-wrap: wrap;
+    justify-content: space-evenly;
+    grid-template-columns: 50% 50%;
+}
+
+.div {
+    display: flex;
+    flex-direction: column;
+    overflow-y: auto;
+    height: 80vh;
+}
+
+.li {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+}
+
+.price_p {
+    display: flex;
+    align-items: center;
+}
+
+.tab_container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    align-content: center;
+    gap: 0 0;
+    flex-wrap: nowrap;
+}
+
+.tab {
+    flex-grow: 1;
+}
